Eliminate warnings p7

+++ demos/include/samples/text_demo/test_text_demo.h
@@ -81,7 +81,7 @@ private:
         };
         
         for (size_t i = 0; i < colors.size(); ++i) {
-            textPainter_->drawText(50, 120 + i * 40, colorNames[i], colors[i]);
+            textPainter_->drawText(50, static_cast<unsigned int>(120 + i * 40), colorNames[i], colors[i]);
         }
     }
     
@@ -103,7 +103,7 @@ private:
         };
         
         for (size_t i = 0; i < sampleText.size(); ++i) {
-            textPainter_->drawText(400, 150 + i * 30, sampleText[i], textColor);
+            textPainter_->drawText(400, static_cast<unsigned int>(150 + i * 30), sampleText[i], textColor);
         }
     }
     
@@ -116,13 +116,13 @@ private:
         
         // Draw border rectangle
         for (int x = 0; x < size.x; ++x) {
-            painter.putPixel(x, 0, borderColor);
-            painter.putPixel(x, size.y - 1, borderColor);
+            painter.putPixel(static_cast<unsigned int>(x), 0, borderColor);
+            painter.putPixel(static_cast<unsigned int>(x), static_cast<unsigned int>(size.y - 1), borderColor);
         }
         
         for (int y = 0; y < size.y; ++y) {
-            painter.putPixel(0, y, borderColor);
-            painter.putPixel(size.x - 1, y, borderColor);
+            painter.putPixel(0, static_cast<unsigned int>(y), borderColor);
+            painter.putPixel(static_cast<unsigned int>(size.x - 1), static_cast<unsigned int>(y), borderColor);
         }
     }
